requirements.txt
Django==1.5.1
dj-database-url==0.2.1
dj-static==0.0.5
gunicorn==17.5
psycopg2==2.5.1
static==0.4
djangorestframework==2.3.6
